pyproject.toml
[project]
name = "ai-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "faiss-cpu>=1.13.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-openai>=0.3.35",
    "openai>=2.16.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.50.0",
]
